+++ src/kr/ac/jbnu/se/tetris/Main.java
@@ -2,6 +2,8 @@ package kr.ac.jbnu.se.tetris;
 
 import java.sql.SQLException;
 
+import kr.ac.jbnu.se.tetris.ui.Select;
+
 public class Main {
     public static void main(String[] args) throws SQLException {
         // LoginPage login = new LoginPage();
