--- src/kr/ac/jbnu/se/tetris/SqlTable.java
@@ -1,5 +1,0 @@
-package kr.ac.jbnu.se.tetris;
-
-public enum SqlTable {
-    NULL,USER_ID, USER_PW
-}
