--- src/kr/ac/jbnu/se/tetris/Tetrominoes.java
@@ -1,5 +1,0 @@
-package kr.ac.jbnu.se.tetris;
-
-public enum Tetrominoes { 
-    NoShape, ZShape, SShape, LineShape, TShape, SquareShape, LShape, MirroredLShape, BombBlock ,LockBlock
-}
