+++ src/kr/ac/jbnu/se/tetris/server/ServerInfo.java
@@ -1,0 +1,8 @@
+package kr.ac.jbnu.se.tetris.server;
+
+public class ServerInfo {
+    public static final String SERVER_URL = "jdbc:mysql://tetris.cjrrqj3pg3pd.ap-northeast-2.rds.amazonaws.com/user";
+    public static final String MANAGEMENT_USER = "tetris";
+    public static final String MANAGEMENT_PASSWORD = "@thvmxmdnpdj";
+    public static final String DRIVER = "com.mysql.cj.jdbc.Driver";
+}
