+++ src/kr/ac/jbnu/se/tetris/server/SqlTable.java
@@ -1,0 +1,5 @@
+package kr.ac.jbnu.se.tetris.server;
+
+public enum SqlTable {
+    NULL,USER_ID, USER_PW
+}
