+++ src/kr/ac/jbnu/se/tetris/shape/Tetrominoes.java
@@ -1,0 +1,5 @@
+package kr.ac.jbnu.se.tetris.shape;
+
+public enum Tetrominoes { 
+    NoShape, ZShape, SShape, LineShape, TShape, SquareShape, LShape, MirroredLShape, BombBlock ,LockBlock
+}
