+++ src/kr/ac/jbnu/se/tetris/ui/Tutorial.java
@@ -1,0 +1,20 @@
+package kr.ac.jbnu.se.tetris.ui;
+import javax.swing.JFrame;
+
+public class Tutorial extends JFrame {
+    
+    private String tutorialpng = "image/tutorials.png";
+    
+    public Tutorial() {
+        setFrame();
+    }
+
+    private void setFrame() {
+        Backgrounds background = new Backgrounds(tutorialpng, 299, 537);
+        add(background.getPane());
+        setSize(299, 537);
+        setLayout(null);
+        setResizable(false);
+        setLocationRelativeTo(null);
+    }
+}
